Change scene and route in one pass
This fixes issues that come up where routes are changed by child
coordinators during the change scene process.

For instance, if A and B are scene coordinators, and A is the parent
of B, if B updates its route on start to go to a specific child, then
we have a problem with the old two step pass. One scenario we saw
looks like this:

route = "b"
A's route set:
  change scene for route "b"
A's change scene:
  set current scene to B
  B.start()
B's start:
  dispatch action to update route to "b/c"
B's route set:
  change scene for subroute "c"
  route current scene for ""
A's original route set:
  route current scene to ""
B's route set:
  change scene for subroute ""

A's original route set has effectively undone the changes that B has made.
For this reason, changing scenes and updating routes should be done in one
pass.