Rework routing and coordinators
Routing now works with animations and with view hierarchies that may need
to be adjusted for arbitrary routing changes to happen, like a modal needing
to be dismissed before another route can be shown.

Support for Swift 2.3 is not yet built out for the changes. Either it will be,
or we'll drop support for it altogether soon.

Routing changes:

- prepareForRoute(_:completionHandler:) has been added to give coordinators
  a chance to clean up the view hierarchy before routing.
- var route { set } is replaced by setRoute(_:completionHandler:).

Coordinator changes:

- SceneCoordinator has been refactored to store a single currentScene
  and tag together, rather than have a separate currentPrefix property.
- SceneCoordinator is now restricted to a single String tag per scene.
- TabBarControllerCoordinator has been restricted to the now concrete
  Cordux.Scene rather than any similar type.
- PresentingCoordinator is restricted to a concrete GeneratingScene type.

Overall the coordinator changes are designed to be less flexible, which
means the framework can lift a little more weight. I haven't seen any other
use cases that go against these choices yet, and they can be revisited if
other use cases arise.