Prevent state changes during subscriber broadcast
Prior to this commit each subscriber was sent the new state via
`self.state` on the store. In the rare case that a subscriber
dispatched an action, this means that the old subscription loop
would dispatch with the new value.

For instance, with two subscribers A and B, if A alters the state
during the notification step, we expect to see the following
behavior:

  Someone dispatches an action
  Store reduces the action to state S
  Store notifies A of S
    A dispatches an action
    Store reduces to state S'
    Store notifies A of S'
    Store notifies B of S'
  Store notifies B of S

The actual behavior here was that after S' was created, S' was used
as the new state even for the outer loop due to looking at store's
state variable.

The store's state still proceeds linearly, so even though B gets
S in the outer loop, the store's state is already set to S'.

Note that the above is very confusing! Dispatching actions during
the new state subscription calls should be used carefully or not
at all!